requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
parameterized==0.8.1
pluggy==1.0.0
pytest==7.2.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.2
requests-mock==1.10.0
responses==0.22.0
six==1.16.0
toml==0.10.2
types-toml==0.10.8.1
urllib3==1.26.14
